Change to libmagic
Switched from the `infer` create to `magic` (rust bindings for `libmagic`). This
should widen the number of formats that get successfully identified, and improve
accuracy. I originally used `infer` because it used a static database of
patterns, unlike `libmagic`, and because I remember having issues with the
`MIME` flag sometimes returning junk in the past. But now that I've tested
things a bit, everything seems to be working correctly. I will do more testing
before releasing `shrink-ray` with `magic`.