Check for symlinks
Added a check at the start of the program that ensures the file is not a
symlink. While this will return an error now, I would like to properly handle
symlinks in the future. Symlinks will have to be resolved, before they are
passed to the conversion program. When output is not given, `shrink-ray` will
replace the original symlink, not the underlying file.